requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-automl==0.1.1
googleapis-common-protos==1.5.3
grpcio==1.15.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
